priority stacked: change activate/deactivate to enabled/disabled
Avoid any confusing between activate(d) and actually being active by
changing the first to enable(d). So enabling a state can make it become
active, but this isn't necessarily the case. While using activate
wording looks closer to making it active. This also avoids confusion
when adding cards to check if a state is "activated" as the wording will
then be "enabled".

Refs #4