single state: support setting active state with a delay
Add support for setting the active state with a delay. When a
delay is already running the currently running delay/activation will be
stopped and only the newly provided state will be set to active after
its own delay has passed.

Feature wise this is meant to be used to for example provide some delay
when deactivating states from the priority stack. As this means there is
the possibility to keep the old state while the actor (user) is
transitioning to the new state.
For example when having a stack consisting of "default",
"using the computer" and "watching TV" where currently only "watching TV"
is active you don't want to immediately switch to "default" when turning
the TV off, while the actor is turning on the computer. So by using a
delay it is possible to keep the existing state, and when the computer
is on the "using the computer" state will be activated immediately
(unless the activation of it has a delay too, but in that case the
state transition will only happen after the "using the computer" delay
is passed).

Fixes #6