add new Set device which can have multiple active states
Having sets where multiple states can be active and there is the
possibility to check if at least one state is active (or none is active)
is usefull to for example combine multiple motion sensors into one
state. So it's easy to for example check 'if any motion sensor is
active' instead of having to check all of them.